+++ src/client/components/Form.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
-import { push } from 'react-router-redux';
+import { push } from 'connected-react-router';
 import { Mutation } from 'react-apollo';
 import formToObj from 'form-data-to-object';