+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2019-02-16
+Fixed bugs with redux routing middleware by upgrading the dependency.
+
+### Changed
+- Upgraded from react-router-redux to connected-router-redux for more stability
+- Cleaned up some old redux routing code
+
 ## [2.1.0] - 2019-02-15
 Improve dynamic routing + optimize route rendering